\name{set_controls_optim}
\alias{set_controls_optim}
\title{Specification of control parameters for \code{optim}}
\usage{
  set_controls_optim(lst)
}
\arguments{
  \item{lst}{a named list}
}
\value{
  a valid list of inputs for optim. See \code{optim} for
  details
}
\description{
  Used to specify options for optim
}

