\name{set_controls_snowfall}
\alias{set_controls_snowfall}
\title{Specification of control parameters for \code{snowfall}}
\usage{
  set_controls_snowfall(lst)
}
\arguments{
  \item{lst}{A list of control parameters}
}
\value{
  A list with named elements "parallel", "cpus"
}
\description{
  Used to specify options for snowfall
}

