\name{sf.control}
\alias{sf.control}
\title{Specification of control parameters for \code{snowfall}}
\usage{
  sf.control(parallel = TRUE, cpus = 2L)
}
\arguments{
  \item{parallel}{logical}

  \item{cpus}{integer}
}
\value{
  A list with named elements "parallel", "cpus"
}
\description{
  Used to specify options for snowfall
}

